pyproject.toml
[project]
name = "redzone"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "asgiref==3.9.1",
    "django==5.2.5",
    "gunicorn==23.0.0",
    "packaging==25.0",
    "pillow==11.3.0",
    "psycopg2-binary==2.9.10",
    "python-decouple==3.8",
    "sqlparse==0.5.3",
    "dj-database-url",
]
